Improve default line height calculation
LineHeight already includes the LineGap. We don't need to add the UnderlinePosition to the line height because the position is always within the LineHeight.

However, we need to add some extra line height because otherwise lines are very close.

Adding an extra 15% of the line height looks good across different font sizes.